+++ RELEASES.md
@@ -8,8 +8,8 @@ Breaking changes
 ----------------
 
 * `AST::iter_functions` now returns an iterator instead of taking a closure.
-* `Module::iter_script_fn_info` is removed and merged into `Module::iter_script_fn`.
 * `Module::get_script_function_by_signature` renamed to `Module::get_script_fn` and returns `&<Shared<ScriptFnDef>>`.
+* `Module::num_fn`, `Module::num_var` and `Module::num_iter` are removed and merged into `Module::count`.
 * The `merge_namespaces` parameter to `Module::eval_ast_as_new` is removed and now defaults to `true`.
 * `GlobalFileModuleResolver` is removed because its performance gain over the `FileModuleResolver` is no longer very significant.
 * The following `EvalAltResult` variants are removed and merged into `EvalAltResult::ErrorMismatchDataType`: `ErrorCharMismatch`, `ErrorNumericIndexExpr`, `ErrorStringIndexExpr`, `ErrorImportExpr`, `ErrorLogicGuard`, `ErrorBooleanArgMismatch`